==> 31_Linear_Regression&Baseline_Model.ipynb <==
### Linear Regression Model
This notebook uses the train and validation model data to train a linear regression model on the training data. I will first use the default parameters in the model and then use optimal parameters generated by a random rearch.

#importing the data
import pandas as pd
import numpy as np
from pathlib import Path
import math
from sklearn.metrics import r2_score

import warnings
warnings.filterwarnings('ignore')

#importing the data
PARENT = "Predicing_House_Prices"
path = Path(PARENT).parent / "../Data/X_train_model2.csv"
X_train_model2 = pd.read_csv(path)

path2 = Path(PARENT).parent / "../Data/X_valid_model2.csv"
X_valid_model2 = pd.read_csv(path2)

path3 = Path(PARENT).parent / "../Data/y_train_model2.csv"
y_train_model2 = pd.read_csv(path3)

path4 = Path(PARENT).parent / "../Data/y_valid_model2.csv"
y_valid_model2 = pd.read_csv(path4)

#dropping unnamed column
X_train_model2 = X_train_model2.drop(columns="Unnamed: 0")
X_valid_model2 = X_valid_model2.drop(columns="Unnamed: 0")
y_train_model2 = y_train_model2.drop(columns="Unnamed: 0")
y_valid_model2 = y_valid_model2.drop(columns="Unnamed: 0")

X_valid_model2.describe()

#training the default linear regression model
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error

lin_reg1 = LinearRegression().fit(X_train_model2, y_train_model2)

#getting the mean squared error and R^2

#y predictions on the validation data
preds= lin_reg1.predict(X_valid_model2)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_valid_model2, preds))))
print("R Squared:", format(lin_reg1.score(X_valid_model2, y_valid_model2)))

### Tuning Parameters

Below I will be getting optimized parameters using random search.

# random search for parameters

from sklearn.model_selection import RandomizedSearchCV
import random

random.seed(3467)
# run randomized search
lin_reg2 = LinearRegression(
    normalize= False,
    n_jobs=None
).fit(X_train_model2, y_train_model2)

#intializing possible parameters
params = {
    "fit_intercept": [True, False],
    "copy_X": [True, False],
    "positive": [True, False]
}
#number of iterations
n_iter_search = 400
random_search = RandomizedSearchCV(
    lin_reg2, param_distributions=params, 
    n_iter=n_iter_search, 
    cv= 3, n_jobs = -1
)
#fitting the random search
random_search.fit(X_train_model2, y_train_model2)
#printing the results
rand_opt = random_search.best_params_
print(rand_opt)

#fitting the model with random search parameters
lin_reg3 = LinearRegression(positive= True, fit_intercept=False, copy_X= True,
    n_jobs=None).fit(X_train_model2, y_train_model2)

#getting the mean squared error and R^2

#y predictions on the validation data
preds2= lin_reg3.predict(X_valid_model2)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_valid_model2, preds2))))
print("R Squared:", format(lin_reg3.score(X_valid_model2, y_valid_model2)))

I will now fit a model with less features. I picked the features that I think would have a strong effect on house price and limited the data to only those features.

# subsetting for only necessary variables
X_train_model3 = X_train_model2.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction", "state_CA", "state_GA"])
X_valid_model3 = X_valid_model2.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction", "state_CA", "state_GA"])


#fitting the model with limited features
lin_reg4 = LinearRegression(positive= True, fit_intercept=False, copy_X= True,
    n_jobs=None).fit(X_train_model3, y_train_model2)

#getting the root mean squared error and R^2

#y predictions on the validation data
preds4= lin_reg4.predict(X_valid_model3)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_valid_model2, preds4))))
print("R Squared:", format(lin_reg4.score(X_valid_model3, y_valid_model2)))

### Baseline Model
The above model, with limited features, will be used as our baseline model so below we will train and test it on the GA, CA subsets, and all data.

#importing the Georgia data
PARENT = "Predicing_House_Prices"
path15 = Path(PARENT).parent / "../Data/X_train_GA.csv"
X_train_GA = pd.read_csv(path15)

path14 = Path(PARENT).parent / "../Data/X_test_GA.csv"
X_test_GA = pd.read_csv(path14)

path13 = Path(PARENT).parent / "../Data/y_train_GA.csv"
y_train_GA = pd.read_csv(path13)

path12 = Path(PARENT).parent / "../Data/y_test_GA.csv"
y_test_GA= pd.read_csv(path12)

path11 = Path(PARENT).parent / "../Data/X_train_GA_noGeo.csv"
X_train_GA_NoGeo = pd.read_csv(path11)

#importing the California data
path5 = Path(PARENT).parent / "../Data/X_train_CA.csv"
X_train_CA = pd.read_csv(path5)

path6 = Path(PARENT).parent / "../Data/X_test_CA.csv"
X_test_CA = pd.read_csv(path6)

path7 = Path(PARENT).parent / "../Data/y_train_CA.csv"
y_train_CA = pd.read_csv(path7)

path8 = Path(PARENT).parent / "../Data/y_test_CA.csv"
y_test_CA= pd.read_csv(path8)

path10 = Path(PARENT).parent / "../Data/X_train_CA_noGeo.csv"
X_train_CA2 = pd.read_csv(path10)

#importing the all train and test data
path16 = Path(PARENT).parent / "../Data/X_train_All.csv"
X_train_all = pd.read_csv(path16)
path17 = Path(PARENT).parent / "../Data/y_train_All.csv"
y_train_all = pd.read_csv(path17)

path18 = Path(PARENT).parent / "../Data/X_test_All.csv"
X_test_all = pd.read_csv(path18)
path19 = Path(PARENT).parent / "../Data/y_test_All.csv"
y_test_all = pd.read_csv(path19)

#dropping unneeded columns from all data
X_train_all_2 = X_train_all.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction", "state_CA", "state_GA"])
X_test_all_2 = X_test_all.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction", "state_CA", "state_GA"])

#fitting baseline model for all train and test data
lin_reg4_All= LinearRegression(positive= True, fit_intercept=False, copy_X= True,
    n_jobs=None).fit(X_train_all_2, y_train_all)

preds_All= lin_reg4_All.predict(X_test_all_2)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_test_all, preds_All))))
print("R Squared:", format(lin_reg4_All.score(X_test_all_2, y_test_all)))

#dropping unnamed column
X_train_GA  = X_train_GA.drop(columns="Unnamed: 0")
X_train_GA_NoGeo = X_train_GA_NoGeo.drop(columns="Unnamed: 0")
X_test_GA  = X_test_GA.drop(columns="Unnamed: 0")
y_train_GA = y_train_GA.drop(columns="Unnamed: 0")
y_test_GA = y_test_GA.drop(columns="Unnamed: 0")

X_train_CA  = X_train_CA .drop(columns="Unnamed: 0")
X_train_CA2 = X_train_CA2.drop(columns="Unnamed: 0")
X_test_CA  = X_test_CA.drop(columns="Unnamed: 0")
y_train_CA = y_train_CA.drop(columns="Unnamed: 0")
y_test_CA = y_test_CA.drop(columns="Unnamed: 0")

#dropping all state, city, and county columns for training GA data
X_train_GA2 = X_train_GA[X_train_GA.columns.drop(list(X_train_GA.filter(regex='state_')))]
X_train_GA3 = X_train_GA2[X_train_GA2.columns.drop(list(X_train_GA2.filter(regex='city_')))]
X_train_GA4 = X_train_GA3[X_train_GA3.columns.drop(list(X_train_GA3.filter(regex='county_')))]
X_train_GA4

#dropping all state, city, and county columns for testing GA data
X_test_GA2 = X_test_GA[X_test_GA.columns.drop(list(X_test_GA.filter(regex='state_')))]
X_test_GA3 = X_test_GA2[X_test_GA2.columns.drop(list(X_test_GA2.filter(regex='city_')))]
X_test_GA4 = X_test_GA3[X_test_GA3.columns.drop(list(X_test_GA3.filter(regex='county_')))]
X_test_GA4

#dropping all state, city, and county columns training CA data
X_train_CA2 = X_train_CA[X_train_CA.columns.drop(list(X_train_CA.filter(regex='state_')))]
X_train_CA3 = X_train_CA2[X_train_CA2.columns.drop(list(X_train_CA2.filter(regex='city_')))]
X_train_CA4 = X_train_CA3[X_train_CA3.columns.drop(list(X_train_CA3.filter(regex='county_')))]
X_train_CA4

#dropping all state, city, and county columns for testing CA data
X_test_CA2 = X_test_CA[X_test_CA.columns.drop(list(X_test_CA.filter(regex='state_')))]
X_test_CA3 = X_test_CA2[X_test_CA2.columns.drop(list(X_test_CA2.filter(regex='city_')))]
X_test_CA4 = X_test_CA3[X_test_CA3.columns.drop(list(X_test_CA3.filter(regex='county_')))]
X_test_CA4

# subsetting for only necessary variables
X_train_GA4_2 = X_train_GA4.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction"])
X_test_GA4_2 = X_test_GA4.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction"])
X_train_CA4_2 = X_train_CA4.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction"])
X_test_CA4_2 = X_test_CA4.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction"])

### Training the Baseline model on the California Data

#training the regression model on only the CA data
lin_reg4_CA = LinearRegression(positive= True, fit_intercept=False, copy_X= True,
    n_jobs=None).fit(X_train_CA4_2, y_train_CA)

#training the regression model on only the CA data
lin_reg4_CA = LinearRegression(positive= True, fit_intercept=False, copy_X= True,
    n_jobs=None).fit(X_train_CA4_2, y_train_CA)

#getting the root mean squared error and R^2

#y predictions on the CA test data
preds4_CA= lin_reg4_CA.predict(X_test_CA4_2)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_test_CA, preds4_CA))))
print("R Squared:", format(lin_reg4_CA.score(X_test_CA4_2, y_test_CA)))

#y predictions on the GA test data
from sklearn.metrics import r2_score
preds_GA_final= lin_reg4_CA.predict(X_test_GA4_2)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_test_GA, preds_GA_final))))
print("R Squared:", format(r2_score(y_test_GA, preds_GA_final)))

### Training the Baseline model on the Georgia Data

#training the regression model on only the GA data
lin_reg_GA = LinearRegression(positive= True, fit_intercept=False, copy_X= True,
    n_jobs=None).fit(X_train_GA4_2, y_train_GA)

#getting the root mean squared error and R^2

#y predictions on the validation data
preds5_GA= lin_reg_GA.predict(X_test_GA4_2)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_test_GA, preds5_GA))))
print("R Squared:", format(r2_score(y_test_GA, preds5_GA)))

#y predictions on the CA test data
preds5_CA= lin_reg_GA.predict(X_test_CA4_2)

print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_test_CA, preds5_CA))))
print("R Squared:", format(lin_reg_GA.score(X_test_CA4_2, y_test_CA)))

### Looking at a model that predicts using the mean price

#getting baseline model of predicting the mean price

mean_price =y_train_model2["price"].mean()
mean_array = np.full(shape=len(y_valid_model2), fill_value=mean_price)
mean_array=mean_array.reshape(len(y_valid_model2),1)
residuals_means = y_valid_model2- mean_array

#getting the root mean squared error and R^2
print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_valid_model2, mean_array))))
print("R Squared:", format(r2_score(y_valid_model2, mean_array)))

#getting baseline model of predicting the mean price for GA data
mean_price_GA =y_train_GA["price"].mean()
mean_array_GA = np.full(shape=len(y_test_GA), fill_value=mean_price_GA)
mean_array_GA=mean_array_GA.reshape(len(y_test_GA),1)


#getting the root mean squared error and R^2
print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_test_GA, mean_array_GA))))
print("R Squared:", format(r2_score(y_test_GA, mean_array_GA)))

#getting baseline model of predicting the mean price for CA data
mean_price_CA =y_train_CA["price"].mean()
mean_array_CA = np.full(shape=len(y_test_CA), fill_value=mean_price_CA)
mean_array_CA=mean_array_CA.reshape(len(y_test_CA),1)


#getting the root mean squared error and R^2
print("Root Mean Squared Error:", format(math.sqrt(mean_squared_error(y_test_CA, mean_array_CA))))
print("R Squared:", format(r2_score(y_test_CA, mean_array_CA)))
