==> 34_model_KNN.ipynb <==
## K-nearest neighbors

Check the performance of the model using KNN.

1. with default setting
2. with optimized hyperparamters using grid search
3. Baseline model with default setting
4. Baseline model with optimized hyperparamters using grid search 

import pandas as pd
import numpy as np
from pathlib import Path
import warnings
warnings.filterwarnings('ignore')

# import data
PARENT = "Predicting_House_Prices"
path = Path(PARENT).parent / "../Data/X_train_model2.csv"
X_train = pd.read_csv(path)

path2 = Path(PARENT).parent / "../Data/y_train_model2.csv"
y_train = pd.read_csv(path2)

path3 = Path(PARENT).parent / "../Data/X_valid_model2.csv"
X_valid = pd.read_csv(path3)

path4 = Path(PARENT).parent / "../Data/y_valid_model2.csv"
y_valid = pd.read_csv(path4)

X_train = X_train.drop(columns="Unnamed: 0")
y_train = y_train.drop(columns="Unnamed: 0")
X_valid = X_valid.drop(columns="Unnamed: 0")
y_valid = y_valid.drop(columns="Unnamed: 0")

X_valid.describe()

print(X_train.shape, X_valid.shape)

### with default setting

from sklearn.neighbors import KNeighborsRegressor
from sklearn.metrics import mean_squared_error

knn = KNeighborsRegressor()
knn.fit(X_train, y_train)

knn_score = knn.score(X_valid, y_valid)
knn_preds = knn.predict(X_valid)

print(
    f" Mean Squared Error : {mean_squared_error(y_valid, knn_preds)} \n"
    f" Score : {knn_score:.5f} \n",
)

(mean_squared_error(y_valid, knn_preds))**0.5

### with optimized hyperparamters using grid search

from sklearn.model_selection import train_test_split

# 30% of training data is used for hyperparameter tuning for efficient search
X_temp, _, y_temp, _ = train_test_split(X_train, y_train, random_state=0, test_size=0.7)
print(len(X_temp), len(y_temp))

# List Hyperparameters
grid_params = { 'n_neighbors' : np.arange(1, 16),
                'weights' : ['uniform','distance'],
                'metric' : ['minkowski','euclidean','manhattan'],
                'p': [1,2]}

from sklearn.model_selection import GridSearchCV

# Use GridSearch
gs = GridSearchCV(KNeighborsRegressor(), grid_params, verbose = 1, cv = 3, n_jobs = -1)

g_res = gs.fit(X_temp, y_temp)

# Print The value of best Hyperparameters
g_res.best_params_

# KNN with optimized parameters
knn2 = KNeighborsRegressor(n_neighbors = 9, weights = 'distance', metric = 'minkowski', p = 1)
knn2.fit(X_train, y_train)

knn2_score = knn2.score(X_valid, y_valid)
knn2_preds = knn2.predict(X_valid)

print(
    f" Mean Squared Error : {mean_squared_error(y_valid, knn2_preds)} \n"
    f" Score : {knn2_score:.5f} \n",
)

(mean_squared_error(y_valid, knn2_preds))**0.5

# KNN with optimized parameters explored before changing preprocessing 
knn3 = KNeighborsRegressor(n_neighbors = 40, weights = 'distance', metric = 'minkowski', p = 1, leaf_size = 15)
knn3.fit(X_train, y_train)

knn3_score = knn3.score(X_valid, y_valid)
knn3_preds = knn2.predict(X_valid)

print(
    f" Mean Squared Error : {mean_squared_error(y_valid, knn3_preds)} \n"
    f" Score : {knn3_score:.5f} \n",
)

import pickle

# functions to save / load pickle objects
def save_obj(obj, filename):
    try:
        with open(filename, "wb") as f:
            pickle.dump(obj, f, protocol=pickle.HIGHEST_PROTOCOL)
    except Exception as ex:
        print("Error:", ex)


def load_obj(filename):
    try:
        with open(filename, "rb") as f:
            return pickle.load(f)
    except Exception as ex:
        print("Error:", ex)

# save model
save_obj(knn2, '../Code/KNN.pickle')

### Baseline model with default setting

# subsetting for only necessary vairables
X_train_sub = X_train.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
                            "pool", "isNewConstruction", "state_CA", "state_GA"])
X_valid_sub = X_valid.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
                            "pool", "isNewConstruction", "state_CA", "state_GA"])

# KNN with default setting
knn4 = KNeighborsRegressor()
knn4.fit(X_train_sub, y_train)

knn4_score = knn4.score(X_valid_sub, y_valid)
knn4_preds = knn2.predict(X_valid)

print(
    f" Mean Squared Error : {mean_squared_error(y_valid, knn4_preds)} \n"
    f" Score : {knn4_score:.5f} \n",
)

### Baseline model with optimized hyperparamters using grid search

# KNN with optimized parameters
knn5 = KNeighborsRegressor(n_neighbors = 14, weights = 'distance', metric = 'minkowski', p = 2)
knn5.fit(X_train_sub, y_train)

knn5_score = knn5.score(X_valid_sub, y_valid)
knn5_preds = knn2.predict(X_valid)

print(
    f" Mean Squared Error : {mean_squared_error(y_valid, knn5_preds)} \n"
    f" Score : {knn5_score:.5f} \n",
)
