==> 35_model_MLP_Classifier.ipynb <==
# **Neural Network**

In the process of selecting a model
taking Neural Network to the test to see whether it works better with our data. 

# Importing Packages

import pandas as pd
import numpy as np
from scipy.stats import loguniform
from sklearn.metrics import mean_squared_error
from sklearn.neural_network import MLPRegressor
from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import RandomizedSearchCV


# Reading Trained and Validation Files


#importing the data
from pathlib import Path
PARENT = "Predicing_House_Prices"
path = Path(PARENT).parent / "../Data/X_train_model2.csv"
X_train = pd.read_csv(path)

path2 = Path(PARENT).parent / "../Data/X_valid_model2.csv"
X_valid = pd.read_csv(path2)

path3 = Path(PARENT).parent / "../Data/y_train_model2.csv"
y_train = pd.read_csv(path3)

path4 = Path(PARENT).parent / "../Data/y_valid_model2.csv"
y_valid = pd.read_csv(path4)

# Checking y_train features
X_train

# Checking length of X_train
len(X_train)

## Dropping Unamed Columns for the Data

# Dropping Unamed Columns
X_train = X_train.drop(columns="Unnamed: 0")
X_valid = X_valid.drop(columns="Unnamed: 0")
y_train = y_train.drop(columns="Unnamed: 0")
y_valid = y_valid.drop(columns="Unnamed: 0")

X_valid.describe()

# Checking Null Values
X_train.isnull().values.any()

# Neural Network with manually tuned parameter's.

# Fitting X_train and y_train to MLp Regressor
mlp =  MLPRegressor(learning_rate_init = 0.03,
                    hidden_layer_sizes = (30,30),
                    solver = 'adam',
                    tol = 1e-5,
                    activation = 'relu',
                    n_iter_no_change = 500,
                    batch_size = 50,random_state=0,max_iter=500).fit(X_train, y_train)


# Predict through X_Validation
predict_y= mlp.predict(X_valid)

predict_y

# Getting Our RMSE and R Squared
print("The Mean Squared Error is %.2f"%(np.sqrt(mean_squared_error(y_valid, predict_y))))
print("R Squared %.3f" % (mlp.score(X_valid, y_valid)))

# Filtering out the feature to see if there is improvement in scores.

# Dropping unused features
X_train1 = X_train.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction", "state_CA", "state_GA"])
X_valid1 = X_valid.filter(["livingArea", "bathrooms", "bedrooms", "garageSpaces", 
"pool", "isNewConstruction", "state_CA", "state_GA"])



# Applying manual parameter tuning with useful data
mlp2=MLPRegressor(learning_rate_init = 0.3,
                    hidden_layer_sizes = (20,),
                    alpha = 0.0001,
                    solver = 'adam',
                    tol = 1e-5,
                    early_stopping = False,
                    activation = 'relu',
                    n_iter_no_change = 100,
                    batch_size = 500,random_state=0,max_iter=5000).fit(X_train1, y_train)

# Predict through new filtered X_Validation
predict_y2= mlp2.predict(X_valid1)

predict_y2

# Getting Our RMSE and R Squared
print("The Mean Squared Error is %.2f "%(np.sqrt(mean_squared_error(y_valid, predict_y2))))
print("R Squared %.5f" % (mlp2.score(X_valid1, y_valid)))

# Trying to use Grid search and Random Search to get optimal parameter

The cell had to be interrupted since it had taken 24 hours to run and was still operating had to go with alternative option which is manually tuned parameters. 

# Trying to use Grid search and get optimal parameter
mlp_gs = MLPRegressor(max_iter=500)
parameter_space = {
    'hidden_layer_sizes': [(10,30,10),(20,)],
    'activation': ['tanh', 'relu'],
    'solver': ['sgd', 'adam'],
    'alpha': [0.0001, 0.05],
    'learning_rate': ['constant','adaptive'],
}

clf = GridSearchCV(mlp_gs, parameter_space, n_jobs=-1, cv=5)
clf.fit(X_train, y_train) # X is train samples and y is the corresponding labels

# Trying to use Random search and get optimal parameter
classifier = MLPRegressor(solver = 'sgd',tol = 1e-5,early_stopping = False,activation = 'relu',n_iter_no_change = 1000,random_state=1,max_iter=1000)
distributions = {'learning_rate_init':loguniform(0.0001,1),'hidden_layer_sizes':[(2,2), (5,5), (30,30)],'alpha':loguniform(1e-8,1e2),'batch_size':[1,3,5,10,20,50,100,250,500]}
randomclf = RandomizedSearchCV(classifier, distributions)
randomclf.fit(X_train, y_train)
randomclf.best_params_
