==> 52_Validating_On_California_Model.ipynb <==
### California Model
This notebook verifies the model on Georgia and California data after it was only trained on California data. 

# Importing Packages
import pandas as pd
import numpy as np
from pathlib import Path
from lightgbm.sklearn import LGBMRegressor
from sklearn.metrics import r2_score
from sklearn.metrics import mean_squared_error as rmse_score


# Importing data for both Georgia and California Data

# Importing the California data

X_train_CA = pd.read_csv("X_train_CA.csv")
X_test_CA = pd.read_csv("X_test_CA.csv")
y_train_CA = pd.read_csv("y_train_CA.csv")
y_test_CA = pd.read_csv("y_test_CA.csv")
X_train_CA2 = pd.read_csv("X_train_CA_noGeo.csv")


X_train_CA2


# Importing the Georgia data
X_train_GA = pd.read_csv("X_train_GA.csv")
X_test_GA = pd.read_csv("X_test_GA.csv")
y_train_GA = pd.read_csv("y_train_GA.csv")
y_test_GA = pd.read_csv("y_test_GA.csv")
X_train_GA_NoGeo = pd.read_csv("X_train_GA_noGeo.csv")


# Dropping Unnamed Columns  for both Georgia and California Data

# Dropping unnamed column
X_train_GA = X_train_GA.drop(columns="Unnamed: 0")
X_train_GA_NoGeo = X_train_GA_NoGeo.drop(columns="Unnamed: 0")
X_test_GA = X_test_GA.drop(columns="Unnamed: 0")
y_train_GA = y_train_GA.drop(columns="Unnamed: 0")
y_test_GA = y_test_GA.drop(columns="Unnamed: 0")

X_train_CA = X_train_CA.drop(columns="Unnamed: 0")
X_train_CA_Geo = X_train_CA2.drop(columns="Unnamed: 0")
X_test_CA = X_test_CA.drop(columns="Unnamed: 0")
y_train_CA = y_train_CA.drop(columns="Unnamed: 0")
y_test_CA = y_test_CA.drop(columns="Unnamed: 0")


# Removing state, county, and city for both Georgia and California Data

# Dropping all state, city, and county columns for training GA data
X_train_GA2 = X_train_GA[
    X_train_GA.columns.drop(list(X_train_GA.filter(regex="state_")))
]
X_train_GA3 = X_train_GA2[
    X_train_GA2.columns.drop(list(X_train_GA2.filter(regex="city_")))
]
X_train_GA4 = X_train_GA3[
    X_train_GA3.columns.drop(list(X_train_GA3.filter(regex="county_")))
]
X_train_GA4


# Dropping all state, city, and county columns for testing GA data
X_test_GA2 = X_test_GA[X_test_GA.columns.drop(list(X_test_GA.filter(regex="state_")))]
X_test_GA3 = X_test_GA2[X_test_GA2.columns.drop(list(X_test_GA2.filter(regex="city_")))]
X_test_GA4 = X_test_GA3[
    X_test_GA3.columns.drop(list(X_test_GA3.filter(regex="county_")))
]
X_test_GA4


# Dropping all state, city, and county columns training CA data
X_train_CA2 = X_train_CA[
    X_train_CA.columns.drop(list(X_train_CA.filter(regex="state_")))
]
X_train_CA3 = X_train_CA2[
    X_train_CA2.columns.drop(list(X_train_CA2.filter(regex="city_")))
]
X_train_CA4 = X_train_CA3[
    X_train_CA3.columns.drop(list(X_train_CA3.filter(regex="county_")))
]
X_train_CA4


# Dropping all state, city, and county columns for testing CA data
X_test_CA2 = X_test_CA[X_test_CA.columns.drop(list(X_test_CA.filter(regex="state_")))]
X_test_CA3 = X_test_CA2[X_test_CA2.columns.drop(list(X_test_CA2.filter(regex="city_")))]
X_test_CA4 = X_test_CA3[
    X_test_CA3.columns.drop(list(X_test_CA3.filter(regex="county_")))
]
X_test_CA4


# Using the model for California only, Validating for both Georgia and California Data

# This function reads in the model's pickle file
def load_obj(filename):
    try:
        with open(filename, "rb") as f:
            return pickle.load(f)
    except Exception as ex:
        print("Error:", ex)


#Installing Pickle5 inorder to read the model
!pip3 install pickle5

# Reading in the California model
import pickle5 as pickle

lgbm_using_CA = load_obj("lgbm_using_CA.pickle")


# Training the model on the California data
lgbm_using_CA.fit(X_train_CA4, y_train_CA)


# Testing the model on the California Data
pred11 = lgbm_using_CA.predict(X_test_CA4)
print("Root Mean Squared Error:", format(np.sqrt(rmse_score(y_test_CA, pred11))))

print("R Squared:", format(r2_score(y_test_CA, lgbm_using_CA.predict(X_test_CA4))))


# Testing the model on the Georgia Data

preds12 = lgbm_using_CA.predict(X_test_GA4)

print("Root Mean Squared Error:", format(np.sqrt(rmse_score(y_test_GA, preds12))))

print("R Squared:", format(r2_score(y_test_GA, lgbm_using_CA.predict(X_test_GA4))))


# We assumed there were baseline differences in pricing between the states since there was a large negative in R Square while testing the model on Georgia data.
As a result, we tried the following method:

Calculate the average price difference between CA and GA:


* Using the CA model as a guide,
Train using CA data, predict with GA data, and then add the mean difference to the projected price to get the "adjusted predicted price."

* Predicted price + mean price difference between the two states = adjusted predicted price

* Using the updated anticipated price, calculate r2 and RMSE. 

# Getting mean price difference between the two states
adjusted = y_train_CA["price"].mean() - y_train_GA["price"].mean()
adjusted


# Adjusted predicted price
new_GA_test = y_test_GA + adjusted


new_GA_test


# Testing the model on the Georgia Data

preds13 = lgbm_using_CA.predict(X_test_GA4)
print("Root Mean Squared Error:", format(np.sqrt(rmse_score(new_GA_test, preds13))))
print("R Squared:", format(r2_score(new_GA_test, lgbm_using_CA.predict(X_test_GA4))))
