==> 53_Validating_On_Georgia_Model.ipynb <==
### Georgia Model
This notebook uses the model that is only trained to the Georgia data and validates it on the California and Georgia data.

#importing Packages
import pandas as pd
import numpy as np
from pathlib import Path
from lightgbm.sklearn import LGBMRegressor
from sklearn.metrics import r2_score
from sklearn.metrics import mean_squared_error as rmse_score 

# Importing data for both Georgia and California Data

#Importing the Georgia data
X_train_GA= pd.read_csv("X_train_GA.csv")
X_test_GA = pd.read_csv("X_test_GA.csv")
y_train_GA = pd.read_csv("y_train_GA.csv")
y_test_GA= pd.read_csv("y_test_GA.csv")
X_train_GA_NoGeo = pd.read_csv("X_train_GA_noGeo.csv")

X_train_GA_NoGeo

#Importing the California data

X_train_CA = pd.read_csv("X_train_CA.csv")
X_test_CA = pd.read_csv("X_test_CA.csv")
y_train_CA = pd.read_csv("y_train_CA.csv")
y_test_CA= pd.read_csv("y_test_CA.csv")
X_train_CA2 = pd.read_csv("X_train_CA_noGeo.csv")

X_train_CA2

# Dropping Unnamed Columns  for both Georgia and California Data

#Dropping unnamed column
X_train_GA  = X_train_GA.drop(columns="Unnamed: 0")
X_train_GA_NoGeo = X_train_GA_NoGeo.drop(columns="Unnamed: 0")
X_test_GA  = X_test_GA.drop(columns="Unnamed: 0")
y_train_GA = y_train_GA.drop(columns="Unnamed: 0")
y_test_GA = y_test_GA.drop(columns="Unnamed: 0")

X_train_CA  = X_train_CA .drop(columns="Unnamed: 0")
X_train_CA_Geo = X_train_CA2.drop(columns="Unnamed: 0")
X_test_CA  = X_test_CA.drop(columns="Unnamed: 0")
y_train_CA = y_train_CA.drop(columns="Unnamed: 0")
y_test_CA = y_test_CA.drop(columns="Unnamed: 0")

# Removing state, county, and city for both Georgia and California Data

#Dropping all state, city, and county columns for training GA data
X_train_GA2 = X_train_GA[X_train_GA.columns.drop(list(X_train_GA.filter(regex='state_')))]
X_train_GA3 = X_train_GA2[X_train_GA2.columns.drop(list(X_train_GA2.filter(regex='city_')))]
X_train_GA4 = X_train_GA3[X_train_GA3.columns.drop(list(X_train_GA3.filter(regex='county_')))]
X_train_GA4

#Dropping all state, city, and county columns for testing GA data
X_test_GA2 = X_test_GA[X_test_GA.columns.drop(list(X_test_GA.filter(regex='state_')))]
X_test_GA3 = X_test_GA2[X_test_GA2.columns.drop(list(X_test_GA2.filter(regex='city_')))]
X_test_GA4 = X_test_GA3[X_test_GA3.columns.drop(list(X_test_GA3.filter(regex='county_')))]
X_test_GA4

#Dropping all state, city, and county columns training CA data
X_train_CA2 = X_train_CA[X_train_CA.columns.drop(list(X_train_CA.filter(regex='state_')))]
X_train_CA3 = X_train_CA2[X_train_CA2.columns.drop(list(X_train_CA2.filter(regex='city_')))]
X_train_CA4 = X_train_CA3[X_train_CA3.columns.drop(list(X_train_CA3.filter(regex='county_')))]
X_train_CA4

#Dropping all state, city, and county columns for testing CA data
X_test_CA2 = X_test_CA[X_test_CA.columns.drop(list(X_test_CA.filter(regex='state_')))]
X_test_CA3 = X_test_CA2[X_test_CA2.columns.drop(list(X_test_CA2.filter(regex='city_')))]
X_test_CA4 = X_test_CA3[X_test_CA3.columns.drop(list(X_test_CA3.filter(regex='county_')))]
X_test_CA4

# Using the model for Georgia only, Validating for both Georgia and California Data

#This function reads in the model's pickle file
def load_obj(filename):
    try:
        with open(filename, "rb") as f:
            return pickle.load(f)
    except Exception as ex:
        print("Error:", ex)

#Installing Pickle5 inorder to read the model
!pip3 install pickle5

#Reading in the Georgia model
import pickle5 as pickle
lgbm_using_GA = load_obj('lgbm_using_GA.pickle')

#Training the model on the GA data
lgbm_using_GA.fit(X_train_GA_NoGeo, y_train_GA)

#Testing the model on the Georgia Data
preds= lgbm_using_GA.predict(X_test_GA4)

print("Root Mean Squared Error:", format(rmse_score(y_test_GA, preds)))

print("R Squared:", format(r2_score(y_test_GA,lgbm_using_GA.predict(X_test_GA4))))


#Testing the model on the California Data
preds2= lgbm_using_GA.predict(X_test_CA4)

print("Root Mean Squared Error:", format(rmse_score(y_test_CA, preds2)))

print("R Squared:", format(r2_score(y_test_CA,lgbm_using_GA.predict(X_test_CA4))))
