==> 91_utility_load_data_and_models.ipynb <==
# import packages
import pandas as pd
import numpy as np
import pickle  # for saving and loading models
from lightgbm.sklearn import LGBMRegressor  # for Light GBM
from sklearn.metrics import r2_score  # for R squared
from sklearn.metrics import mean_squared_error as rmse_score  # for root mean squared error

## Load Models
- The final models are:
  - lgbm_all_data
  - lgbm_using_GA
  - lgbm_using_CA

# functions to save / load pickle objects
def save_obj(obj, filename):
    try:
        with open(filename, "wb") as f:
            pickle.dump(obj, f, protocol=pickle.HIGHEST_PROTOCOL)
    except Exception as ex:
        print("Error:", ex)


def load_obj(filename):
    try:
        with open(filename, "rb") as f:
            return pickle.load(f)
    except Exception as ex:
        print("Error:", ex)


# Model trained using all data
lgbm_all_data = load_obj('../models/lgbm_all_data.pickle')

# Model trained using only data from Georgia
lgbm_using_GA = load_obj('../models/lgbm_using_GA.pickle')

# Model trained using only data from California
lgbm_using_CA = load_obj('../models/lgbm_using_CA.pickle')


## Load Data
- I modified the original X_train, X_validate, X_test, etc. a little
- The original data sets produced errors when training LGBM
- Use the data sets below for LGBM

# I'm including these codes for convenience
# It's totally possible that you might not need all these data sets for your part of the validation
# Change directory path if needed

# Data with state, county, city info
# Use these data sets with lgbm_all_data
X_train = pd.read_csv('../large_data_files/Data_for_presentation/X_train_for_LGBM.csv')
X_train_all=pd.read_csv('../large_data_files/Data_for_presentation/X_train_all_for_LGBM.csv')
X_test=pd.read_csv('../large_data_files/Data_for_presentation/X_test_for_LGBM.csv')

# Data without state, county, city info
# Use these data sets with lgbm_using_CA or lgbm_using_GA
X_test_noGeo=pd.read_csv('../large_data_files/Data_for_presentation/X_test_noGeo.csv')
X_train_all_noGeo=pd.read_csv('../large_data_files/Data_for_presentation/X_train_all_noGeo.csv')
X_train_GA=pd.read_csv('../large_data_files/Data_for_presentation/X_train_GA_noGeo.csv')
X_train_CA=pd.read_csv('../large_data_files/Data_for_presentation/X_train_CA_noGeo.csv')
